pyproject.toml
[project]
name = "desktop-automation-botcity"
version = "1.0.0"
description = "Vision-Based Desktop Automation with Dynamic Icon Grounding using BotCity"
requires-python = ">=3.8"
dependencies = [
    "botcity-framework-core>=1.0.0",
    "botcity-maestro-sdk>=1.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
